Move ErrorBoundary and Loadable components to client/app/components
